Integrated configs into items, allowed 1.16.4

+++ src/main/java/syric/dragonseeker/item/tool/dragonseekerGeneric.java
@@ -48,7 +48,7 @@ public class dragonseekerGeneric extends Item {
 //        );
 //    }
 
-    public dragonseekerGeneric(int opDistIn, int maxDistIn, double minPingIn, double maxPingIn, int minSigIn, double powIn, float minVolIn, float maxVolIn, float minPitchIn, float maxPitchIn, SoundEvent negSoundIn, SoundEvent pingSoundIn, boolean detectsCorpsesIn, boolean detectsTameIn, int durabilityIn, Rarity rarityIn, Item repairItemIn, int seekerTypeIn) {
+    public dragonseekerGeneric(int opDistIn, int maxDistIn, double minPingIn, double maxPingIn, int minSigIn, double powIn, double minVolIn, double maxVolIn, double minPitchIn, double maxPitchIn, SoundEvent negSoundIn, SoundEvent pingSoundIn, boolean detectsCorpsesIn, boolean detectsTameIn, int durabilityIn, Rarity rarityIn, Item repairItemIn, int seekerTypeIn) {
         super(new Properties()
                 .stacksTo(1)
                 .tab(IceAndFire.TAB_ITEMS)
@@ -61,10 +61,10 @@ public class dragonseekerGeneric extends Item {
         maxPing = maxPingIn;
         minSig = minSigIn;
         pow = powIn;
-        minVol = minVolIn;
-        maxVol = maxVolIn;
-        minPitch = minPitchIn;
-        maxPitch = maxPitchIn;
+        minVol = (float) minVolIn;
+        maxVol = (float) maxVolIn;
+        minPitch = (float) minPitchIn;
+        maxPitch = (float) maxPitchIn;
         negSound = negSoundIn;
         pingSound = pingSoundIn;
         detectsCorpses = detectsCorpsesIn;
